Modal fixed with correct coins and co2
